Add account health service

+++ src/main/java/org/willy/crypto/connexion/coinbase/objects/price/PriceCB.java
@@ -3,6 +3,8 @@ package org.willy.crypto.connexion.coinbase.objects.price;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,5 +13,5 @@ import lombok.experimental.FieldDefaults;
 public class PriceCB {
     String base;
     String currency;
-    Float amount;
+    BigDecimal amount;
 }
